Hybris: move code to the proper files
