+++ libs/hybris/src/hooks_net.c
@@ -1,9 +1,11 @@
 #include <stdlib.h>
 #include <sys/socket.h>
 #include <netdb.h>
+#include "hooks_net.h"
 
 #include "../include/hybris/hook.h"
 
+#ifndef __APPLE__
 struct android_addrinfo {
     int	ai_flags;	/* AI_PASSIVE, AI_CANONNAME, AI_NUMERICHOST */
     int	ai_family;	/* PF_xxx */
@@ -13,11 +15,12 @@ struct android_addrinfo {
     char	*ai_canonname;	/* canonical name for hostname */
     struct	sockaddr *ai_addr;	/* binary address */
     struct	android_addrinfo *ai_next;	/* next structure in linked list */
-
-    struct	addrinfo *_real_addrinfo;
 };
-struct android_addrinfo* convert_addrinfo(struct addrinfo* res) {
-    struct android_addrinfo* ares = (struct android_addrinfo*) malloc(sizeof(struct android_addrinfo));
+
+struct android_addrinfo* convert_addrinfo(struct addrinfo* res)
+{
+    struct android_addrinfo* ares = (struct android_addrinfo*) malloc(
+            sizeof(struct android_addrinfo) + sizeof(struct addrinfo *));
     ares->ai_flags = res->ai_flags;
     ares->ai_family = res->ai_family;
     ares->ai_socktype = res->ai_socktype;
@@ -29,6 +32,7 @@ struct android_addrinfo* convert_addrinfo(struct addrinfo* res) {
     if (res->ai_next != NULL) {
         ares->ai_next = convert_addrinfo(res->ai_next);
     }
+    *((struct addrinfo **) (ares + 1)) = res;
     return ares;
 }
 
@@ -52,7 +56,6 @@ int my_getaddrinfo(const char *node, const char *service,
     }
     if (res != NULL) {
         *ares = convert_addrinfo(res);
-        (*ares)->_real_addrinfo = res;
     } else {
         *ares = NULL;
     }
@@ -60,7 +63,7 @@ int my_getaddrinfo(const char *node, const char *service,
 }
 
 void my_freeaddrinfo(struct android_addrinfo *ai) {
-    freeaddrinfo(ai->_real_addrinfo);
+    freeaddrinfo(*((struct addrinfo **) (ai + 1)));
     struct android_addrinfo *ai_next;
     while (ai) {
         ai_next = ai->ai_next;
@@ -68,6 +71,7 @@ void my_freeaddrinfo(struct android_addrinfo *ai) {
         ai = ai_next;
     }
 }
+#endif
 
 #define	A_NI_NOFQDN	0x00000001
 #define	A_NI_NUMERICHOST	0x00000002
@@ -75,10 +79,7 @@ void my_freeaddrinfo(struct android_addrinfo *ai) {
 #define	A_NI_NUMERICSERV	0x00000008
 #define	A_NI_DGRAM	0x00000010
 
-int my_getnameinfo (const struct sockaddr *__restrict sa,
-                    socklen_t salen, char *__restrict host,
-                    socklen_t hostlen, char *__restrict serv,
-                    socklen_t servlen, int flags) {
+int convert_getnameinfo_flags(int flags) {
     int glibc_flags = 0;
     if (flags & A_NI_NOFQDN)
         glibc_flags |= A_NI_NOFQDN;
@@ -90,20 +91,55 @@ int my_getnameinfo (const struct sockaddr *__restrict sa,
         glibc_flags |= NI_NUMERICSERV;
     if (flags & A_NI_DGRAM)
         glibc_flags |= NI_DGRAM;
+    return glibc_flags;
+}
+
+#ifndef __APPLE__
+int my_getnameinfo (const struct sockaddr *__restrict sa,
+                    socklen_t salen, char *__restrict host,
+                    socklen_t hostlen, char *__restrict serv,
+                    socklen_t servlen, int flags) {
+    int glibc_flags = convert_getnameinfo_flags(flags);
     return getnameinfo(sa, salen, host, hostlen, serv, servlen, glibc_flags);
 }
+#endif
 
 
 static struct _hook net_hooks[] = {
     /* net specifics, to avoid __res_get_state */
+#ifndef __APPLE__
     {"getaddrinfo", my_getaddrinfo},
     {"freeaddrinfo", my_freeaddrinfo},
+    {"getnameinfo", my_getnameinfo},
+#endif
     {"gethostbyaddr", gethostbyaddr},
     {"gethostbyname", gethostbyname},
     {"gethostbyname2", gethostbyname2},
     {"gethostent", gethostent},
-    {"getnameinfo", my_getnameinfo},
     {"gai_strerror", gai_strerror},
+    /* socket.h */
+#ifndef __APPLE__
+    {"socket", socket},
+    {"bind", bind},
+    {"sendto", sendto},
+    {"recvfrom", recvfrom},
+    {"getsockname", getsockname},
+    {"getsockopt", getsockopt},
+    {"setsockopt", setsockopt},
+#endif
+    {"socketpair", socketpair},
+    {"connect", connect},
+    {"getpeername", getpeername},
+    {"send", send},
+    {"recv", recv},
+    {"sendmsg", sendmsg},
+    // {"sendmmsg", sendmmsg},
+    {"recvmsg", recvmsg},
+    // {"recvmmsg", recvmmsg},
+    {"listen", listen},
+    {"accept", accept},
+    // {"accept4", accept4},
+    {"shutdown", shutdown},
     {NULL, NULL}
 };
 REGISTER_HOOKS(net_hooks)
