+++ libs/hybris/src/hooks_net.h
@@ -1,0 +1,6 @@
+#ifndef HOOKS_NET_H
+#define HOOKS_NET_H
+
+int convert_getnameinfo_flags(int flags);
+
+#endif /* HOOKS_NET_H */
